requirements.txt
Django==4.0.4
gunicorn==20.0.4
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
Deprecated==1.2.13
django-auditlog==1.0.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-jsonfield-backport==1.0.5
django-modeladmin-reorder==0.3.1
django-oauth-toolkit==1.7.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jsonfield==3.1.0
jwcrypto==1.0
Markdown==3.3.6
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
pycparser==2.21
pyparsing==3.0.8
python-dateutil==2.6.0
pytz==2022.1
requests==2.27.1
rest-condition==1.0.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
